Unsigned int

+++ linux-5.4.9/kernel/state_changes.c
@@ -20,7 +20,7 @@ SYSCALL_DEFINE2(print_state_changes, pid_t, pid, int, milliseconds)
 
   struct state_change *state_changes;
   u64 time_ns = ktime_get();
-  u64 max_time = time_ns - milliseconds * 1000000;
+  u64 max_time = time_ns - (((unsigned int)milliseconds) * 1000000);
 
   printk("CURR: %ld\n", time_ns);
   printk("MAX: %ld\n", max_time);
